+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Added `addSegmentationToViewWithID:` in view interface for adding segmentation to an ongoing view
 * Added `addSegmentationToViewWithName:` in view interface for adding segmentation to an ongoing view
 
+* Fixed pause view recursive call issue
+
 ## 23.8.3
 * Added `requestDropAgeHours` initial config property to set a time limit after which the requests would be removed if not sent to the server
 * Added a call to enroll users to A/B tests when getting a remote config value: 'getValueAndEnroll'